move tests and test data
