+++ projects/_03_independent_cqc/_08_diagnostics/jobs/diagnostics_on_known_filled_posts.py
@@ -7,7 +7,9 @@ from utils.column_names.ind_cqc_pipeline_columns import (
     IndCqcColumns as IndCQC,
     PartitionKeys as Keys,
 )
-from utils.diagnostics_utils import diagnostics_utils as dUtils
+from projects._03_independent_cqc._08_diagnostics.utils import (
+    diagnostics_utils as dUtils,
+)
 
 partition_keys = [Keys.year, Keys.month, Keys.day, Keys.import_date]
 estimate_filled_posts_columns: list = [
