+++ projects/_03_independent_cqc/_08_diagnostics/tests/utils/test_diagnostics_utils.py
@@ -3,18 +3,22 @@ from unittest.mock import patch, Mock
 
 from pyspark.sql import WindowSpec
 
-from jobs.diagnostics_on_known_filled_posts import (
+from projects._03_independent_cqc._08_diagnostics.jobs.diagnostics_on_known_filled_posts import (
     absolute_value_cutoff,
     standardised_value_cutoff,
     percentage_value_cutoff,
 )
-import utils.diagnostics_utils.diagnostics_utils as job
-from tests.test_file_schemas import DiagnosticsUtilsSchemas as Schemas
-from tests.test_file_data import DiagnosticsUtilsData as Data
+import projects._03_independent_cqc._08_diagnostics.utils.diagnostics_utils as job
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_schemas import (
+    DiagnosticsUtilsSchemas as Schemas,
+)
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_data import (
+    DiagnosticsUtilsData as Data,
+)
 from utils import utils
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
 
-PATCH_PATH: str = "utils.diagnostics_utils.diagnostics_utils"
+PATCH_PATH: str = "projects._03_independent_cqc._08_diagnostics.utils.diagnostics_utils"
 
 
 class DiagnosticsUtilsTests(unittest.TestCase):
